py/nrf_mesh/requirements.txt
paho-mqtt
pyserial
requests
tzlocal
